added environments, platform service, authorization with claimReq for roles

+++ src/app/shared/services/auth.service.ts
@@ -2,14 +2,20 @@ import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Router } from '@angular/router';
 import { LocalStorageKeys } from '../constants/constants';
+import { PlatformService } from './platform.service';
+import { environment } from '../../../environments/environment';
 
 @Injectable({
   providedIn: 'root',
 })
 export class AuthService {
-  private baseUrl = 'https://localhost:7102/api';
+  private baseUrl = environment.apiBaseUrl;
 
-  constructor(private httpClient: HttpClient, private router: Router) {}
+  constructor(
+    private httpClient: HttpClient,
+    private router: Router,
+    private platformService: PlatformService
+  ) {}
 
   // ** Register the user ** //
   registerUser(formData: any) {
@@ -30,32 +36,63 @@ export class AuthService {
 
   // ** Save tokens to local storage ** //
   saveTokensToLocalStorage(tokens: any) {
-    if (tokens.accessToken) {
-      localStorage.setItem(LocalStorageKeys.ACCESS_TOKEN, tokens.accessToken);
-    }
-    if (tokens.refreshToken) {
-      localStorage.setItem(LocalStorageKeys.REFRESH_TOKEN, tokens.refreshToken);
+    if (this.platformService.isBrowser()) {
+      if (tokens.accessToken) {
+        localStorage.setItem(LocalStorageKeys.ACCESS_TOKEN, tokens.accessToken);
+      }
+      if (tokens.refreshToken) {
+        localStorage.setItem(
+          LocalStorageKeys.REFRESH_TOKEN,
+          tokens.refreshToken
+        );
+      }
     }
   }
 
   // ** Get access token from local storage ** //
   getAccessToken() {
-    return localStorage.getItem(LocalStorageKeys.ACCESS_TOKEN) || '';
+    if (this.platformService.isBrowser()) {
+      return localStorage.getItem(LocalStorageKeys.ACCESS_TOKEN) || '';
+    } else {
+      return '';
+    }
   }
 
   // ** Get refresh token from local storage ** //
   getRefreshToken() {
-    return localStorage.getItem(LocalStorageKeys.REFRESH_TOKEN) || '';
+    if (this.platformService.isBrowser()) {
+      return localStorage.getItem(LocalStorageKeys.REFRESH_TOKEN) || '';
+    } else {
+      return '';
+    }
   }
 
   // ** Log out the user ** //
   logOutUser() {
-    localStorage.clear();
+    if (this.platformService.isBrowser()) {
+      localStorage.clear();
+    }
     this.router.navigateByUrl('/sign-in');
   }
 
   // ** Check if the user is logged in ** //
   isUserLoggedIn() {
     return !!this.getAccessToken() && !!this.getRefreshToken();
+  }
+
+  // ** Get the user claims ** //
+  getUserClaims() {
+    if (this.platformService.isBrowser()) {
+      const accessToken = this.getAccessToken();
+      if (accessToken) {
+        const payload = accessToken.split('.')[1];
+        const decodedPayload = atob(payload);
+        return JSON.parse(decodedPayload);
+      } else {
+        return null;
+      }
+    } else {
+      return null;
+    }
   }
 }
